requirements.txt
# Core AI and Browser Automation
browser-use>=0.1.0
langchain-openai>=0.1.0
openai>=1.0.0

# Web Framework and Real-time Communication
Flask>=2.3.0
Flask-SocketIO>=5.3.0
python-socketio>=5.8.0
eventlet>=0.33.0

# Database and Data Processing
psycopg2-binary>=2.9.0
pandas>=2.0.0
python-dotenv>=1.0.0

# PDF and File Handling
requests>=2.31.0

# Web Browser Automation (Playwright backend)
playwright>=1.35.0

# Additional Utilities
argparse  # Built-in, but explicit for clarity 